Added interfaces in another file and imported them in user.model.ts
