+++ Backend/src/types/user.types.ts
@@ -1,0 +1,10 @@
+// src/types/user.types.ts
+
+import { Document } from 'mongoose';
+
+export interface IUser extends Document {
+  name?: string;
+  email: string;
+  password: string;
+  comparePassword(candidatePassword: string): Promise<boolean>;
+}
